- jbj: backport assorted rpmio changes from HEAD to stay in sync. - jbj: make sure the open path is set when using [bgl]zdOpen vector too. - jbj: lzdio: rework the layering to be more similar to gzdio/bzdio. - jbj: rpmrepo: fix: --lzma segfault, lzdFdopen was not saving path. - jbj: splint: annotate argvAppend differently. - jbj: add _fdopen to FDIO_t, expose fdbg(), make lzdio.c standalone. - jbj: expose XfdLink/XfdFree/XfdNew, nuke the FDIO_t vectors. - jbj: eliminate unused FDIO_t vectors. XfdLink/XfdFree/XfdNew need nuking. - jbj: create lzdio.c to carry lzma comprssion, possibly w dlopen(). - jbj: add --enable-build-gcov for CFLAGS -fprofile-arcs -ftest-coverage.

+++ rpmio/rpmssl.c
@@ -172,11 +172,11 @@ int rpmsslVerifyRSA(pgpDig dig)
     xx = ll;	/* WRONG WRONG WRONG */
     dbuf = xcalloc(1, ll);
     /* XXX FIXME: what parameter goes into dbuf? */
-    while (xx < ll)
+    while (xx < (int)ll)
 	memmove(&dbuf[1], dbuf, xx++), dbuf[0] = 0;
     xx = RSA_public_decrypt(ll, dbuf, dbuf, ssl->rsa, RSA_PKCS1_PADDING);
 /*@=moduncon@*/
-    rc = (xx == nb && (memcmp(rsahm, dbuf, nb) == 0));
+    rc = (xx == (int)nb && (memcmp(rsahm, dbuf, nb) == 0));
     dbuf = _free(dbuf);
     rsahm = _free(rsahm);
